# docs/swagger.yaml
basePath: /
definitions:
  article.CreateInput:
    properties:
      content:
        type: string
      is_published:
        type: boolean
      media:
        items:
          $ref: '#/definitions/article.MediaInput'
        type: array
      slug:
        maxLength: 255
        minLength: 2
        type: string
      tests:
        items:
          $ref: '#/definitions/article.TestLinkInput'
        type: array
      title:
        maxLength: 255
        minLength: 2
        type: string
    required:
    - title
    type: object
  article.Media:
    properties:
      article_id:
        type: integer
      caption:
        type: string
      id:
        type: integer
      media_type:
        type: string
      sort_order:
        type: integer
      url:
        type: string
    type: object
  article.MediaInput:
    properties:
      caption:
        type: string
      media_type:
        enum:
        - image
        - video
        type: string
      sort_order:
        type: integer
      url:
        type: string
    required:
    - media_type
    - url
    type: object
  article.Model:
    properties:
      author_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_published:
        type: boolean
      media:
        items:
          $ref: '#/definitions/article.Media'
        type: array
      slug:
        type: string
      tests:
        items:
          $ref: '#/definitions/article.TestLink'
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
  article.TestLink:
    properties:
      article_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      test_id:
        type: integer
    type: object
  article.TestLinkInput:
    properties:
      description:
        type: string
      test_id:
        type: integer
    required:
    - test_id
    type: object
  article.UpdateInput:
    properties:
      content:
        type: string
      is_published:
        type: boolean
      media:
        items:
          $ref: '#/definitions/article.MediaInput'
        type: array
      slug:
        maxLength: 255
        minLength: 2
        type: string
      tests:
        items:
          $ref: '#/definitions/article.TestLinkInput'
        type: array
      title:
        maxLength: 255
        minLength: 2
        type: string
    type: object
  auth.AuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/user.Model'
    type: object
  auth.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  auth.ForgotPasswordInput:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  auth.LoginInput:
    properties:
      email:
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  auth.MessageResponse:
    properties:
      message:
        type: string
    type: object
  auth.RefreshInput:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.RegisterInput:
    properties:
      email:
        type: string
      first_name:
        maxLength: 100
        minLength: 2
        type: string
      last_name:
        maxLength: 100
        minLength: 2
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      phone:
        maxLength: 20
        minLength: 10
        type: string
    required:
    - email
    - first_name
    - last_name
    - name
    - password
    - phone
    type: object
  auth.ResendVerificationInput:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  auth.ResetPasswordInput:
    properties:
      new_password:
        maxLength: 72
        minLength: 8
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
  auth.ResetTokenResponse:
    properties:
      expires_at:
        type: string
      token:
        type: string
    type: object
  auth.TokenPair:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  auth.VerifyEmailInput:
    properties:
      token:
        type: string
    required:
    - token
    type: object
...
  user.Model:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified_at:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: Документация API сервиса it_courses.
  title: API it_courses
  version: "1.0"
paths:
...
  /articles:
    get:
      description: Возвращает список всех образовательных статей
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/article.Model'
            type: array
      summary: Список статей
      tags:
      - articles
    post:
      consumes:
      - application/json
      description: Создаёт новую образовательную статью. Доступно teacher и admin
      parameters:
      - description: Данные статьи
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/article.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/article.Model'
        "400":
          description: Bad Request
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
      summary: Создать статью
      tags:
      - articles
  /articles/{id}:
    delete:
      description: Удаляет статью по ID. Доступно teacher и admin
      parameters:
      - description: ID статьи
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            type: string
      summary: Удалить статью
      tags:
      - articles
    put:
      consumes:
      - application/json
      description: Обновляет данные статьи. Доступно teacher и admin
      parameters:
      - description: ID статьи
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/article.UpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/article.Model'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
      summary: Обновить статью
      tags:
      - articles
  /articles/{slug}:
    get:
      description: Возвращает статью с медиа и ссылками на тесты по slug
      parameters:
      - description: Slug статьи
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/article.Model'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Детали статьи
      tags:
      - articles
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Отправляет письмо со ссылкой для сброса пароля
      parameters:
      - description: Email пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.ForgotPasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.ResetTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Запрос восстановления пароля
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Аутентификация по email и паролю, возвращает JWT токены
      parameters:
      - description: Учётные данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Вход в систему
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Получает новую пару access + refresh токенов по refresh токену
      parameters:
      - description: Refresh токен
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TokenPair'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Обновление JWT токена
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя и отправляет письмо для подтверждения
        email
      parameters:
      - description: Данные для регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Регистрация пользователя
      tags:
      - auth
  /auth/resend-verification:
    post:
      consumes:
      - application/json
      description: Генерирует новый токен и отправляет письмо подтверждения
      parameters:
      - description: Email пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.ResendVerificationInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Повторная отправка подтверждения email
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Устанавливает новый пароль по токену из письма восстановления
      parameters:
      - description: Токен и новый пароль
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.ResetPasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Сброс пароля
      tags:
      - auth
  /auth/verify-email:
    post:
      consumes:
      - application/json
      description: Подтверждает email пользователя по токену из письма
      parameters:
      - description: Токен подтверждения
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.VerifyEmailInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Подтверждение email
      tags:
      - auth
...
swagger: "2.0"